+++ src/zope/configuration/config.py
@@ -333,6 +333,16 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
     includepath = ()
     info = ''
 
+    #: These `Exception` subclasses are allowed to be raised from `execute_actions`
+    #: without being re-wrapped into a `ConfigurationExecutionError`. (`BaseException`
+    #: instances are never wrapped.)
+    #:
+    #: Users of instances of this class may modify this before calling `execute_actions`
+    #: if they need to propagate specific exceptions.
+    #:
+    #: .. versionadded:: 4.2.0
+    pass_through_exceptions = ()
+
     def __init__(self):
         super(ConfigurationMachine, self).__init__()
         self.actions = []
@@ -367,6 +377,9 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
 
         This calls the action callables after resolving conflicts.
         """
+        pass_through_exceptions = self.pass_through_exceptions
+        if testing:
+            pass_through_exceptions = BaseException
         try:
             for action in resolveConflicts(self.actions):
                 callable = action['callable']
@@ -377,11 +390,9 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
                 info = action['info']
                 try:
                     callable(*args, **kw)
-                except (KeyboardInterrupt, SystemExit): # pragma: no cover
+                except pass_through_exceptions:
                     raise
-                except:
-                    if testing:
-                        raise
+                except Exception:
                     t, v, tb = sys.exc_info()
                     try:
                         reraise(ConfigurationExecutionError(t, v, info),